pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "outlier-x"
version = "0.1.0"
description = "Production-grade Python CLI tool for sports betting data ingestion, normalization, and validation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Outlier.bet Team", email = "dev@outlier.bet"}
]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "click>=8.0.0,<9.0.0",
    "pyyaml>=6.0,<7.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests>=2.28.0,<3.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "pyarrow>=14.0.0,<15.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "sphinx>=6.0.0",
]

[project.scripts]
outlier = "src.main:cli"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
